feat: Sign Windows executables
Initial changes to support signing Windows executables before they are attached to the GitHub release.

Because Windows signing has to take place on a Windows machine, this requires the build to be split into different phases. First the binaries are built, then the Windows executables are Authenticode signed, then the signed files are attached to the GitHub release.

This may require further modularisation to allow the files that go inside the installer to also be signed, depending on how flexible the `sign` tool is to recurse into the installer to sign the contents (or not).