feat: Refactor Windows signing
Refactor into a multi-stage pipeline so that the binaries are built, signed, then the installer is built again with those files, then everything collated at the end for publishing.

This approach currently does not sign the uninstaller as we need to run build the installer on a Windows image to be able to intercept the creation of the uninstaller to sign it before it is added to the installer.

Signed-off-by: martincostello <martin@martincostello.com>